# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for simpo pharmacy management system supporting POS, inventory, reporting, and multi-branch operations.",
        "title": "simpo Pharmacy Management System API",
        "termsOfService": "https://simpo.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@simpo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/backups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns list of all available backups with metadata (Story 6.3, AC5)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backups"
                ],
                "summary": "List all backups",
                "responses": {
                    "200": {
                        "description": "Backups retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.BackupListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers a database backup operation (Story 6.3, AC5)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backups"
                ],
                "summary": "Trigger manual backup",
                "parameters": [
                    {
                        "description": "Backup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Backup started",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBackupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Backup already in progress",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/backups/cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually triggers cleanup of backups older than retention period (Story 6.3, AC3)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backups"
                ],
                "summary": "Cleanup old backups",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "30",
                        "name": "retention_days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cleanup completed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/backups/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns current backup configuration (Story 6.3, AC8)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backups"
                ],
                "summary": "Get backup configuration",
                "responses": {
                    "200": {
                        "description": "Backup configuration",
                        "schema": {
                            "$ref": "#/definitions/dto.BackupConfig"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates backup configuration settings (Story 6.3, AC8)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backups"
                ],
                "summary": "Update backup configuration",
                "parameters": [
                    {
                        "description": "Backup configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BackupConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/backups/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns current backup job status and statistics (Story 6.3, AC4)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Backups"
                ],
                "summary": "Get backup status",
                "responses": {
                    "200": {
                        "description": "Backup status retrieved",
                        "schema": {
                            "$ref": "#/definitions/dto.BackupJobStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BackupConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "retention_days": {
                    "type": "integer",
                    "example": 30
                },
                "schedule": {
                    "description": "Cron expression",
                    "type": "string",
                    "example": "0 2 * * *"
                },
                "storage_path": {
                    "type": "string",
                    "example": "/backups"
                }
            }
        },
...
        "dto.BackupListResponse": {
            "type": "object",
            "properties": {
                "backups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BackupInfo"
                    }
                },
                "last_backup": {
                    "$ref": "#/definitions/dto.BackupInfo"
                },
                "retention_days": {
                    "type": "integer"
                },
                "total_size": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateBackupRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Manual backup before system update"
                }
            }
        },
        "dto.CreateBackupResponse": {
            "type": "object",
            "properties": {
                "estimated_time": {
                    "type": "string",
                    "example": "2-5 min"
                },
                "filename": {
                    "type": "string",
                    "example": "simpo_20260527_143000.dump"
                },
                "message": {
                    "type": "string",
                    "example": "Backup operation started"
                },
                "status": {
                    "type": "string",
                    "example": "started"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}